# trading-api-service/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "OKX HFT Trading API Service",
    "description": "REST & SSE streaming API for the OKX High-Frequency Trading demo.\nPowered by QuestDB time-series database with real-time OKX market data ingestion.\n\nAll data is sourced exclusively from **QuestDB via R2DBC** — the feed handler service\n(okx-feed-handler) is the single ingestion point from OKX WebSocket. This service is\na **pure read layer** and never connects to OKX directly.\n\n## Transport\n\n- All endpoints are served over **HTTPS/H2** on port `8443`.\n- TLS uses an **ECDSA P-256** leaf certificate issued by a local mkcert CA — see the project README for setup.\n- **HTTP/2** multiplexes all SSE streams over a single TCP connection, eliminating the browser's 6-connection-per-host limit that plagued HTTP/1.1 SSE.\n\n## SSE Streaming\n\nAll `/api/stream/*` endpoints return `text/event-stream` responses.\n\n- **Immediate first poll**: data arrives within milliseconds of subscribing (no initial delay).\n- **Heartbeat keepalive**: a `:heartbeat` SSE comment is sent every 15s to keep the connection alive.\n- **`:connected` comment**: sent immediately on subscribe to flush HTTP response headers.\n- **Delta-polling**: tick-level streams (trades, orderbook, spread, candles) poll QuestDB every 500ms; indicator streams (RSI, Bollinger, VWAP) poll every 3s.\n- **Graceful shutdown**: all streams terminate cleanly on server shutdown.",
    "version": "2.0.0",
    "contact": {
...
    }
  },
  "servers": [
...
  "paths": {
...
    "/api/orderbook": {
      "get": {
        "operationId": "getOrderBook",
        "summary": "Get latest order book",
        "description": "Returns the latest top-5 bid/ask order book snapshot using QuestDB LATEST ON. Returns a single object (not an array).",
        "tags": ["Order Book"],
        "parameters": [
          { "$ref": "#/components/parameters/Symbol" }
        ],
        "responses": {
          "200": {
            "description": "Latest order book snapshot (single object)",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/OrderBook" }
              }
            }
          }
        }
      }
    },
    "/api/orderbook/spread": {
      "get": {
        "operationId": "getSpreadHistory",
        "summary": "Get spread history",
        "description": "Returns bid-ask spread history aggregated using QuestDB SAMPLE BY over the last hour.",
        "tags": ["Order Book"],
        "parameters": [
          { "$ref": "#/components/parameters/Symbol" },
          {
            "name": "interval",
            "in": "query",
            "required": false,
            "description": "Aggregation interval for spread sampling",
            "schema": { "type": "string", "default": "5s", "enum": ["1s", "5s", "15s", "1m"] }
          }
        ],
        "responses": {
          "200": {
            "description": "List of spread snapshots",
            "content": {
              "application/json": {
                "schema": { "type": "array", "items": { "$ref": "#/components/schemas/Spread" } }
              }
            }
          }
        }
      }
    },
    "/api/indicators/vwap": {
      "get": {
        "operationId": "getVwap",
        "summary": "Get VWAP",
        "description": "Returns Volume Weighted Average Price computed using QuestDB CUMULATIVE window function over the last hour in 5s buckets.",
        "tags": ["Indicators"],
        "parameters": [
          { "$ref": "#/components/parameters/Symbol" },
          { "$ref": "#/components/parameters/Limit" }
        ],
        "responses": {
          "200": {
            "description": "Array of up to 500 rows, ordered by ts descending",
            "content": {
              "application/json": {
                "schema": { "type": "array", "items": { "$ref": "#/components/schemas/VwapPoint" } },
                "example": [
                  { "ts": "2026-03-26T15:34:35Z", "close": 69144.9, "vwap": 69034.20 }
                ]
              }
            }
          }
        }
      }
    },
    "/api/indicators/bollinger": {
      "get": {
        "operationId": "getBollingerBands",
        "summary": "Get Bollinger Bands",
        "description": "Returns 20-period Bollinger Bands (SMA ± 2σ) computed using QuestDB window functions over the last hour in 5s buckets.",
        "tags": ["Indicators"],
        "parameters": [
          { "$ref": "#/components/parameters/Symbol" },
          { "$ref": "#/components/parameters/Limit" }
        ],
        "responses": {
          "200": {
            "description": "Array of up to 500 rows, ordered by ts descending",
            "content": {
              "application/json": {
                "schema": { "type": "array", "items": { "$ref": "#/components/schemas/BollingerBandPoint" } },
                "example": [
                  { "ts": "2026-03-26T15:34:35Z", "close": 69144.9, "sma20": 68994.375, "upperBand": 69069.50, "lowerBand": 68919.25 }
                ]
              }
            }
          }
        }
      }
    },
    "/api/indicators/rsi": {
      "get": {
        "operationId": "getRsi",
        "summary": "Get RSI (14-period)",
        "description": "Returns 14-period Relative Strength Index (SMA-smoothed) computed using QuestDB window functions with LAG over the last hour in 5s buckets.",
        "tags": ["Indicators"],
        "parameters": [
          { "$ref": "#/components/parameters/Symbol" },
          { "$ref": "#/components/parameters/Limit" }
        ],
        "responses": {
          "200": {
            "description": "Array of up to 500 rows, ordered by ts descending",
            "content": {
              "application/json": {
                "schema": { "type": "array", "items": { "$ref": "#/components/schemas/RsiPoint" } },
                "example": [
                  { "ts": "2026-03-26T15:34:35Z", "close": 69144.9, "rsi": 32.89 }
                ]
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "parameters": {
      "Symbol": {
        "name": "symbol",
        "in": "query",
        "required": false,
        "description": "Trading pair instrument ID (e.g. BTC-USDT, ETH-USDT, SOL-USDT)",
        "schema": {
          "type": "string",
          "default": "BTC-USDT",
          "examples": ["BTC-USDT", "ETH-USDT", "SOL-USDT"]
        }
      },
      "Limit": {
        "name": "limit",
        "in": "query",
        "required": false,
        "description": "Max data points to return (1–500)",
        "schema": { "type": "integer", "default": 500, "minimum": 1, "maximum": 500 }
      },
...
    },
    "schemas": {
...
      "OrderBook": {
        "type": "object",
        "required": ["ts", "symbol"],
        "properties": {
          "ts":        { "type": "string", "format": "date-time", "description": "Snapshot timestamp" },
          "symbol":    { "type": "string", "description": "Instrument ID" },
          "bid1Price": { "type": "number", "description": "Best bid price (level 1)" },
          "bid1Size":  { "type": "number", "description": "Best bid size (level 1)" },
          "bid2Price": { "type": "number" }, "bid2Size": { "type": "number" },
          "bid3Price": { "type": "number" }, "bid3Size": { "type": "number" },
          "bid4Price": { "type": "number" }, "bid4Size": { "type": "number" },
          "bid5Price": { "type": "number" }, "bid5Size": { "type": "number" },
          "ask1Price": { "type": "number", "description": "Best ask price (level 1)" },
          "ask1Size":  { "type": "number", "description": "Best ask size (level 1)" },
          "ask2Price": { "type": "number" }, "ask2Size": { "type": "number" },
          "ask3Price": { "type": "number" }, "ask3Size": { "type": "number" },
          "ask4Price": { "type": "number" }, "ask4Size": { "type": "number" },
          "ask5Price": { "type": "number" }, "ask5Size": { "type": "number" },
          "midPrice":  { "type": "number", "description": "Mid-market price: (bid1 + ask1) / 2" },
          "spread":    { "type": "number", "description": "Bid-ask spread: ask1 - bid1" }
        }
      },
      "Spread": {
        "type": "object",
        "required": ["ts", "symbol", "bestBid", "bestAsk", "spread", "midPrice"],
        "properties": {
          "ts":       { "type": "string", "format": "date-time", "description": "Bucket timestamp" },
          "symbol":   { "type": "string", "description": "Instrument ID" },
          "bestBid":  { "type": "number", "description": "Best bid price" },
          "bestAsk":  { "type": "number", "description": "Best ask price" },
          "spread":   { "type": "number", "description": "Bid-ask spread" },
          "midPrice": { "type": "number", "description": "Mid price" }
        }
      },
      "VwapPoint": {
        "type": "object",
        "required": ["ts", "close", "vwap"],
        "properties": {
          "ts":    { "type": "string", "format": "date-time" },
          "close": { "type": "number", "description": "Closing price of the 5s candle" },
          "vwap":  { "type": "number", "description": "Cumulative volume-weighted average price" }
        }
      },
      "BollingerBandPoint": {
        "type": "object",
        "required": ["ts", "close", "sma20", "upperBand", "lowerBand"],
        "properties": {
          "ts":        { "type": "string", "format": "date-time" },
          "close":     { "type": "number", "description": "Closing price of the 5s candle" },
          "sma20":     { "type": "number", "description": "20-period simple moving average" },
          "upperBand": { "type": "number", "description": "SMA20 + 2σ" },
          "lowerBand": { "type": "number", "description": "SMA20 − 2σ" }
        }
      },
      "RsiPoint": {
        "type": "object",
        "required": ["ts", "close", "rsi"],
        "properties": {
          "ts":    { "type": "string", "format": "date-time" },
          "close": { "type": "number", "description": "Closing price of the 5s candle" },
          "rsi":   { "type": "number", "minimum": 0, "maximum": 100, "description": "14-period RSI value" }
        }
      },
...
    }
  }
}